+++ app/javascript/components/StaffDashboard/index.js
@@ -3,7 +3,7 @@ import InputBase from '@material-ui/core/InputBase';
 import IconButton from '@material-ui/core/IconButton';
 import SearchIcon from '@material-ui/icons/Search';
 
-import ParticipantCard from './ParticipantCard';
+import ParticipantCard from 'components/ParticipantCard';
 import Navbar from 'components/Navbar';
 import PropTypes from 'prop-types';
 import { Grid } from '@material-ui/core';
